updating delete route

+++ controllers/user.js
@@ -29,10 +29,24 @@ router.get('/unsubscribe', (req, res) => {
 //Delete User
 router.delete('/confirmation', (req, res) => {
     let email = req.body.email
-    User.findOneAndDelete({email})
-        .then(() => {
-            console.log('deleted')
-            res.render('Unsubscribed', {email})
+    let found
+    // check if user exists
+    User.find({email})
+        .then((user) => {
+            // explains no user found
+            if(!user){
+                found = false
+                console.log('Not deleted')
+                res.render('Unsubscribed', {found, email})
+            } else {
+                // delete existing user
+                found = true
+                User.findOneAndDelete({email})
+                    .then(() => {
+                        console.log('deleted')
+                        res.render('Unsubscribed', {email, found})
+                    })
+            }
         })
         .catch((err) => {
             res.status(400).send(err)
